+++ components/datetime/benches/datetime.rs
@@ -6,7 +6,7 @@ mod fixtures;
 use criterion::{criterion_group, criterion_main, Criterion};
 use std::fmt::Write;
 
-use icu_datetime::date::MockDateTime;
+use icu_datetime::mock::MockDateTime;
 use icu_datetime::DateTimeFormat;
 
 fn datetime_benches(c: &mut Criterion) {
