+++ components/datetime/src/provider.rs
@@ -1,6 +1,7 @@
 // This file is part of ICU4X. For terms of use, please see the file
 // called LICENSE at the top level of the ICU4X source tree
 // (online at: https://github.com/unicode-org/icu4x/blob/master/LICENSE ).
+
 use crate::date;
 use crate::error::DateTimeFormatError;
 use crate::fields;
