+++ components/datetime/tests/datetime.rs
@@ -3,7 +3,7 @@
 // (online at: https://github.com/unicode-org/icu4x/blob/master/LICENSE ).
 mod fixtures;
 
-use icu_datetime::date::MockDateTime;
+use icu_datetime::mock::MockDateTime;
 use icu_datetime::DateTimeFormat;
 use std::fmt::Write;
